+++ tests/integration/MediaWiki/Schema/TableCreateReadDeleteTest.php
@@ -104,7 +104,7 @@ class TableCreateReadDeleteTest extends \PHPUnit_Framework_TestCase {
 				new FieldDefinition( 'textfield', new TypeDefinition( TypeDefinition::TYPE_BLOB ), FieldDefinition::NOT_NULL ),
 				new FieldDefinition( 'intfield', new TypeDefinition( TypeDefinition::TYPE_INTEGER ), FieldDefinition::NOT_NULL, 3 ),
 				new FieldDefinition( 'floatfield', new TypeDefinition( TypeDefinition::TYPE_FLOAT ), FieldDefinition::NULL ),
-				new FieldDefinition( 'tinyintfield', new TypeDefinition( TypeDefinition::TYPE_TINYINT ), FieldDefinition::NOT_NULL, true ),
+				new FieldDefinition( 'tinyintfield', new TypeDefinition( TypeDefinition::TYPE_TINYINT ), FieldDefinition::NOT_NULL, 1 ),
 			),
 			array(
 				new IndexDefinition( 'PRIMARY', array( 'intfield' => 0 ), IndexDefinition::TYPE_PRIMARY ),
@@ -143,7 +143,6 @@ class TableCreateReadDeleteTest extends \PHPUnit_Framework_TestCase {
 		);
 
 		$tableReader = $this->newTableReader();
-
 		$this->assertEquals(
 			$table,
 			$tableReader->readDefinition( $table->getName() )